# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Golf Court Reservation API",
        "description": "Documentación de la API para el sistema de reservas de canchas de golf.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/clubs": {
            "get": {
                "tags": [
                    "Clubs"
                ],
                "summary": "Obtener listado de clubs",
                "description": "Retorna una lista paginada de clubs con filtros opcionales por nombre, estado y slug.",
                "operationId": "getClubsList",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Cantidad de elementos por página (máximo 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 25
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filtrar por nombre del club",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por estado del club",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "slug",
                        "in": "query",
                        "description": "Filtrar por slug",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    }
                }
            },
            "post": {
                "tags": [
                    "Clubs"
                ],
                "summary": "Crear un nuevo club",
                "description": "Registra un nuevo club de golf en el sistema.",
                "operationId": "storeClub",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Club de Golf Costa del Sol"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "club-de-golf-costa-del-sol"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "contacto@golfcostasol.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+34952123456"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active"
                                    },
                                    "configuration": {
                                        "type": "object",
                                        "example": {
                                            "courts_count": 9
                                        }
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Carretera de Cádiz Km 204"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Club creado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/clubs/{club}": {
            "get": {
                "tags": [
                    "Clubs"
                ],
                "summary": "Mostrar un club específico",
                "description": "Retorna la información detallada de un club mediante su ID.",
                "operationId": "getClubById",
                "parameters": [
                    {
                        "name": "club",
                        "in": "path",
                        "description": "ID del club",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "404": {
                        "description": "Club no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Clubs"
                ],
                "summary": "Actualizar un club existente",
                "description": "Modifica los datos de un club existente. El campo 'name' es obligatorio.",
                "operationId": "updateClub",
                "parameters": [
                    {
                        "name": "club",
                        "in": "path",
                        "description": "ID del club",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Club Actualizado"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "nuevo@golfcostasol.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+34952999888"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Club actualizado con éxito"
                    },
                    "404": {
                        "description": "Club no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clubs"
                ],
                "summary": "Eliminar un club",
                "description": "Elimina un club del sistema de forma lógica.",
                "operationId": "deleteClub",
                "parameters": [
                    {
                        "name": "club",
                        "in": "path",
                        "description": "ID del club",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Club eliminado exitosamente (sin contenido)"
                    },
                    "404": {
                        "description": "Club no encontrado"
                    }
                }
            }
        },
        "/clubs/{club}/slug": {
            "patch": {
                "tags": [
                    "Clubs"
                ],
                "summary": "Actualizar el slug de un club",
                "description": "Modifica específicamente el slug de un club mediante una petición dedicada.",
                "operationId": "updateClubSlug",
                "parameters": [
                    {
                        "name": "club",
                        "in": "path",
                        "description": "ID del club",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "slug"
                                ],
                                "properties": {
                                    "slug": {
                                        "type": "string",
                                        "example": "nuevo-slug-personalizado"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Slug actualizado con éxito"
                    },
                    "422": {
                        "description": "Error de validación o slug duplicado"
                    }
                }
            }
        },
...
    },
...
}